refactor: simplify ProbabilityPlot component and introduce useChartData for better data handling
